Replace `RoutineAgentSpec` to `RoutineAgentTests` to remove `Quick/Nimble` library
